This version has lots of debugging and almost works; what's missing is the logic to scan forward after there is a collision between the symbol passed in and another declarator - i.e. they both hash to the same symbol table cell.
The solution is to increment the symbol table pointer by one.  But since
the current symbol table is a table of declarator pointers, this doesn't
work.  If you have struct declaration *sp, and it's pointing to some cell
in the symbol table, when you do ++sp it actually increments sp by the size
of a struct declarator.  What we really want is an increment the size of
a struct declarator pointer.

In light of this, and because of the overly indirect nature of dealing
with an array of struct declarator pointers, i'm going to change the
symbol table to be an array of declarators.  It's much cleaner that way.